update design for dashboard page when there is no position
