+++ liquidity/lib/Positions/index.ts
@@ -1,8 +1,5 @@
-export * from './PositionNotConnected';
 export * from './PositionTableHeader';
 export * from './LockedCollateral';
 export * from './PositionsRow';
 export * from './PositionsList';
-export * from './PositionEmpty';
 export * from './TableDivider';
-export * from './PositionsRowLoading';
