+++ liquidity/lib/Positions/package.json
@@ -18,9 +18,7 @@
     "@snx-v3/useCollateralTypes": "workspace:*",
     "@snx-v3/useIsAndromedaStataUSDC": "workspace:*",
     "@snx-v3/useLiquidityPosition": "workspace:*",
-    "@snx-v3/useLiquidityPositions": "workspace:*",
     "@snx-v3/useLocks": "workspace:*",
-    "@snx-v3/useParams": "workspace:*",
     "@snx-v3/useWithdrawTimer": "workspace:*",
     "@synthetixio/wei": "^2.74.4",
     "date-fns": "^2.30.0",
