+++ liquidity/ui/package.json
@@ -40,6 +40,7 @@
     "@snx-v3/useCreateAccount": "workspace:*",
     "@snx-v3/useIsAndromedaStataUSDC": "workspace:*",
     "@snx-v3/useLiquidityPosition": "workspace:*",
+    "@snx-v3/useLiquidityPositions": "workspace:*",
     "@snx-v3/useParams": "workspace:*",
     "@synthetixio/devtools-formatters": "workspace:*",
     "@synthetixio/magic-wallet": "workspace:*",
